Added chunk sub allocations to GPU page Refactor memory reporting Added instance hash to framebuffer to support resizing

+++ GraphicsEngine/Source/Core/Rendering/RayTracing/RHIStateObject.h
@@ -7,8 +7,8 @@ class Camera;
 class ShaderBindingTable;
 struct RHIStateObjectDesc
 {
-	uint64_t AttibuteSize = 0;
-	uint64_t PayloadSize = 0;
+	uint AttibuteSize = 0;
+	uint PayloadSize = 0;
 	// PERFOMANCE TIP: Set max recursion depth as low as needed 
 	// as drivers may apply optimization strategies for low recursion depths.
 	uint MaxRecursionDepth = 1;
