+++ GraphicsEngine/Source/D3D12RHI/RHI/RenderAPIs/D3D12/Raytracing/D3D12HighLevelAccelerationStructure.h
@@ -23,7 +23,7 @@ public:
 	static D3D12_RAYTRACING_ACCELERATION_STRUCTURE_BUILD_FLAGS GetBuildFlags(int BuildFlags);
 private:
 	GPUResource* instanceDescs = nullptr;
-	int LAstCount = 0;
+	size_t LAstCount = 0;
 	GPUResource* scratchSpace = nullptr;
 	D3D12_BUILD_RAYTRACING_ACCELERATION_STRUCTURE_DESC topLevelBuildDesc = {};
 };
